Move StartDedicateadAuthServer back
