+++ cli/commands/service/orchestrator/orchestrator_test.go
@@ -27,6 +27,7 @@ package orchestrator
 
 import (
 	"bytes"
+	"clouditor.io/clouditor/service"
 	"context"
 	"encoding/json"
 	"fmt"
@@ -39,7 +40,6 @@ import (
 	"clouditor.io/clouditor/api/orchestrator"
 	"clouditor.io/clouditor/cli"
 	"clouditor.io/clouditor/cli/commands/login"
-	service_auth "clouditor.io/clouditor/service/auth"
 	service_orchestrator "clouditor.io/clouditor/service/orchestrator"
 
 	"github.com/spf13/viper"
@@ -53,7 +53,6 @@ import (
 var (
 	sock                net.Listener
 	server              *grpc.Server
-	authServices        *service_auth.Service
 	orchestratorService *service_orchestrator.Service
 )
 
@@ -68,10 +67,9 @@ func TestMain(m *testing.M) {
 		panic(err)
 	}
 
-	authServices = service_auth.NewService(service_auth.WithApiKeySaveOnCreate(false))
 	orchestratorService = service_orchestrator.NewService()
 
-	sock, server, err = authServices.StartDedicatedServer(":0")
+	sock, server, _, err = service.StartDedicatedAuthServer(":0")
 	if err != nil {
 		panic(err)
 	}
